+++ src/domain/bc/mod.rs
@@ -7,5 +7,5 @@ pub use periodic::*;
 use crate::util::*;
 
 pub trait BCCheck<const GRID_DIMENSION: usize>: Sync {
-    fn check(&self, world_coord: &Coord<GRID_DIMENSION>) -> Option<f32>;
+    fn check(&self, world_coord: &Coord<GRID_DIMENSION>) -> Option<f64>;
 }
